update dashboard layout
